Streamlined Efficiency
Streamlined the Efficiency of results recording with a switch statement and changed the game loop to a conditions as not to go above 7 wins per player

+++ Darts_501/Darts_501/Player.cpp
@@ -54,7 +54,7 @@ void Player::Add_Win()
 	wins++;
 }
 
-float Player::Get_Wins()
+int Player::Get_Wins()
 {
 	return wins;
 }
